ajout de fonctions: telecharger, effacer, nouveau fichier
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@2671 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ skel_editor/exec/skeleditor.php
@@ -43,6 +43,44 @@ function show_skel_file($path,$current_file) {
   return $output;
 }
 
+// build select menu to choose directory
+function editor_form_directory($path,$depth="") { 
+  $path = "../$path";
+  $folder = dir($path); 
+  if ($depth=="") $output .= "<option value='$path'>(racine)</option>\n"; 
+  $depth .= "&nbsp;&nbsp;&nbsp;&nbsp;";
+  while ($myfile = $folder->read()) {
+      $entirePath  = $path."/".$myfile;        
+      if (substr($myfile,0,1) !=".") {  // exclude hidden files and directories                 
+          if (is_dir($entirePath)) {
+              $output .= "<option value='$entirePath'>$depth$myfile</option>\n";              
+              $output .= editor_form_directory(substr($entirePath,3),$depth); // recursive !           
+          }
+      }      
+  }
+  return $output;  
+}
+
+// add file form
+function editor_addfile() {
+  global $dossier_squelettes;
+  $output = bouton_block_invisible('editor_newfile');
+  $output .= "<img src='../plugins/skel_editor/img_pack/action_add.png' alt='new' />Nouveau fichier";
+  $output .= debut_block_invisible('editor_newfile');
+  $output .= "<form method='get'>\n"; 
+  $output .= "<input type='hidden' name='exec' value='skeleditor' />"; 
+  $output .= "<input type='hidden' name='action' value='new' />"; 
+  //$output .= "nom du fichier:<br />\n";
+  $output .= "<input type='text' name='f' value='untitled.html' onfocus=\"this.value=''\" />"; 
+  $output .= "position:<br />\n"; 
+  $output .= "<select name='target'><br />\n"; 
+  $output .= editor_form_directory($dossier_squelettes);      
+  $output .= "</select><br /><input type='submit' name='sub' value='Cr&eacute;er' />";
+  $output .= "</form>\n";
+  $output .= fin_block();
+  return $output;        
+}
+
 // -------------------------------
 // Main 
 // ------------------------------
@@ -54,7 +92,7 @@ function exec_skeleditor(){
   global $connect_statut;
 	global $connect_toutes_rubriques;
   if ($connect_statut != '0minirezo' OR !$connect_toutes_rubriques) {    
-		debut_page(_T('titre'), "sauver_config", "plugin");
+		debut_page(_T('titre'), "skel_editor", "plugin");
 		echo _T('avis_non_acces_page');
 		fin_page();
 		exit;
@@ -63,20 +101,23 @@ function exec_skeleditor(){
 	// Version de base, a supprimer donc...
 	// globals
 	// $GLOBALS['dossier_squelettes']);
-	
+;
 	// Va chercher le repertoire des squelettes et seulement celui-ci
 	// NB: Ce fichier peut initialiser $dossier_squelettes (old-style)
 	// donc il faut l'inclure "en globals"
-	if ($f = include_spip('mes_fonctions', false)) {
+	if ($fpp = include_spip('mes_fonctions', false)) {
 		global $dossier_squelettes;
-		@include ($f); 
+		@include ($fpp); 
 	}
+	
 	if (@is_readable(_DIR_SESSIONS."charger_plugins_fonctions.php")){
 		// chargement optimise precompile
 		include_once(_DIR_SESSIONS."charger_plugins_fonctions.php");     	
 	}
 	
-	// Action ?
+	// --------------------------------------------------------------------------- 
+	// Action ? 
+	// ---------------------------------------------------------------------------
 	$log = "";
 	if (isset($_POST['editor'])) {      // save file ?
 	     $editor = $_POST['editor'];
@@ -90,25 +131,68 @@ function exec_skeleditor(){
              } else if (fwrite($handle, $editor) === FALSE) {
                  $log = "<span style='color:red'>erreur: impossible d'&ecute;crire dans le fichier</span>";           
              } else {
-                 $log = "<span style='color:green'>fichier sauvegardé @ ".date('H:m')."</span>";
+                 $log = "<span style='color:green'>fichier sauvegardé @ ".date('H:i')."</span>";
                  fclose($handle);
-             }        
-	          
+             }  
+             
        } else {
             $log = "<span style='color:red'>erreur: fichier non éditable en écriture</span>";
        }
   }	 
-
-	// URL request ?
+  
+  // GET request ?
+  $action = "";
 	if (isset($_GET['f'])) {
 	    $file_name = $_GET['f'];
-      $file_name = "..".str_replace("..", "", $file_name);    // security  
+      $file_name = "..".str_replace("..", "", $file_name);    // security 
+            
+      if (isset($_GET['action'])) {
+          $action = $_GET['action'];          
+          if ($action=="delete") {                // delete the file
+            @unlink($file_name);          
+          } else if ($action=="download") {       // download the file
+              if ($file_tmp = @file("$file_name")) {
+                $file_name_nopath = substr($file_name, strrpos($file_name,"/")+1); 
+                $file_str = implode ('',$file_tmp);
+                //header("Content-type: text/plain"); // text/plain or binary .... 
+                header("Content-Disposition: attachment; filename=\"$file_name_nopath\"");
+                echo $file_str; 
+                exit; 
+              }
+          } else if ($action=="new") {            // add new file
+            if (isset($_GET['target'])) {
+               $target = "..".str_replace("..", "", $_GET['target']);    // security                         
+               if (substr($target,3, strlen($dossier_squelettes))== $dossier_squelettes) { // security 2: right place ?                  
+                   // FIXME: check if extension available ?
+                   $file_name = $target."/".substr($file_name,2);
+                   if (is_file($file_name)) { // security 3: overight ?
+                    $log = "<span style='color:red'>erreur: ce fichier existe d&eacute;j&agrave;</span>";
+                   } else {
+                     if (!$handle = fopen($file_name, 'w')) {
+                         $log = "<span style='color:red'>erreur: impossible de cr&eacute;er le fichier. V&eacute;rifier les droits en &eacute;critutre de votre r&eacute;pertoire squelette</span>";
+                     } else if (fwrite($handle, "...") === FALSE) {
+                         $log = "<span style='color:red'>erreur: impossible de cr&eacute;er le fichier. V&eacute;rifier les droits en &eacute;critutre de votre r&eacute;pertoire squelette</span>";           
+                     } else {
+                         $log = "<span style='color:green'>fichier cr&eacute;e @ ".date('H:i')."</span>";
+                         fclose($handle);
+                     }
+                    }
+               }               	
+            }
+           
+             
+              
+          }
+      } 
   } else {
-      $file_name = "";
+      $file_name = "";  
   }
-	
-	//  
-  // HTML output  
+  
+  
+
+  // ---------------------------------------------------------------------------
+  // HTML output 
+  // ---------------------------------------------------------------------------
 	debut_page("Editer le squelette", "naviguer", "plugin");	
   debut_gauche();
 	debut_boite_info();
@@ -116,29 +200,41 @@ function exec_skeleditor(){
 	echo "dossier squelette: <strong>$dossier_squelettes</strong><br />";
 	echo show_skel_file($dossier_squelettes,$file_name);
 	fin_boite_info();
+	echo "<br />";
+	debut_boite_info();
+	echo editor_addfile();
+  fin_boite_info();
 	
 	debut_droite();
 
 	// something to do ?	
-	if ($file_name !="") {   
+	if ($file_name!="") {   
        echo "<div>Fichier: <strong>$file_name</strong> $log</div>\n"; // add extra infos on file:  size ? date ? ...
-       // img or text ?
-       $extension =  strtolower(substr($file_name, strrpos($file_name,".")+1)); 
-       if (in_array($extension,$img_extension)) {     // display file as img
-          echo "<div style='border:1px solid #333;padding:20px;background:#eee'><img src='$file_name' alt='picture' /></div>\n";
-       } else {  // edit file as text  
-          if ($file_tmp = @file("$file_name")) {
-              $file_str = implode ('',$file_tmp);
-              $file_str = str_replace("</textarea","&lt;/textarea",$file_str); // exception: textarea closing tag              
-              echo "<form method='post' action='?exec=skeleditor&amp;retour=skeleditor&amp;f=".urlencode($file_name)."'>\n";
-              echo "<textarea name='editor' cols='80' rows='50'>$file_str</textarea>\n";               
-    	        echo "<input type='submit' name='action' value='Sauver' />";	        
-    	        echo "</form>\n";
-          } else {       
-              echo "<p>Erreur: impossible d'ouvrir ou d'éditer ce fichier.</p>\n";
-          }   
-       }     
-      
+       if ($action=="delete") {
+         echo "<p style='color:green'>Fichier effacé avec succès!</p>\n";
+       } else { 
+           // tools bar
+           echo "<div id='skel_toolbar' style='width:100%;text-align:right;'>\n";
+           echo "<img src='../plugins/skel_editor/img_pack/action_dl.png' alt='download' /><a href=\"?exec=skeleditor&amp;f=".urlencode($file_name)."&amp;action=download\">télécharger</a>";
+           echo "<img src='../plugins/skel_editor/img_pack/action_del.png' alt='delete' /><a href=\"?exec=skeleditor&amp;f=".urlencode($file_name)."&amp;action=delete\" onclick=\"javascript:return confirm('Voulez vous vraiment effacer ce fichier ?');\">effacer</a>";
+           echo "</div>\n";
+           // img or text ?
+           $extension =  strtolower(substr($file_name, strrpos($file_name,".")+1)); 
+           if (in_array($extension,$img_extension)) {     // display file as img
+              echo "<div style='border:1px solid #333;padding:20px;background:#eee'><img src='$file_name' alt='picture' /></div>\n";
+           } else {  // edit file as text  
+              if ($file_tmp = @file("$file_name")) {
+                  $file_str = implode ('',$file_tmp);
+                  $file_str = str_replace("</textarea","&lt;/textarea",$file_str); // exception: textarea closing tag              
+                  echo "<form method='post' action='?exec=skeleditor&amp;retour=skeleditor&amp;f=".urlencode($file_name)."'>\n";
+                  echo "<textarea name='editor' cols='80' rows='50'>$file_str</textarea>\n";               
+        	        echo "<input type='submit' name='action' value='Sauver' />";	        
+        	        echo "</form>\n";
+              } else {
+                  echo "<p>Erreur: impossible d'ouvrir ou d'éditer ce fichier.</p>\n";
+              }   
+           }
+       }
   } else {
       echo "<p>Choississez le fichier que vous voulez &eacute;diter ou visualiser.</p>\n";
   }
